requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastuuid==0.14.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
importlib_metadata==9.0.0
Jinja2==3.1.6
jiter==0.13.0
json_repair==0.58.7
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
litellm==1.83.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
numpy==2.4.4
openai==2.30.0
packaging==26.0
pandas==3.0.2
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.3.32
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
tushare==1.4.29
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websocket-client==1.9.0
yarl==1.23.0
zipp==3.23.0
